feat: Implement Spotify and Trading flows, introduce a new FlowLayout component, and add a Landing page.
