+++ packages/ui/src/lib/flows/spotify/components/Pagination.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-  import { searchOptions, totalTracks } from '@lib/flows/spotify/stores';
-  import { loadTracks } from '@lib/flows/spotify/api';
+  import { searchOptions, totalTracks } from '../stores';
+  import { loadTracks } from '../api';
 
   let currentPage = $derived($searchOptions.page || 1);
   let limit = $derived($searchOptions.limit || 24);
